Refactored controller/service.py and controller/router.py

+++ src/controller/service.py
@@ -6,8 +6,13 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from pymodbus.client.tcp import AsyncModbusTcpClient
 from pymodbus.exceptions import ConnectionException
 
+from auth.models import User
 from controller.models import Controller
 from controller.schemas import ControllerCreateUpdate
+from controller.exceptions import (
+    PermissionForControllerDenied,
+    ControllerNotFound,
+)
 from database import get_async_session
 
 
@@ -20,16 +25,25 @@ class ControllerService:
         result = await self.session.execute(stmt)
         controller = result.scalars().first()
 
-        if controller is not None:
-            controller = controller.to_dict()
+        if controller is None:
+            raise ControllerNotFound
 
-        return controller
+        return controller.to_dict()
 
-    async def read_controller_data(self, controller_id: int, data_address: int = 0, data_length: int = 1):
+    async def read_controller_data(
+            self,
+            user_id: int,
+            controller_id: int,
+            data_address: int = 0,
+            data_length: int = 1
+    ):
         controller = await self.get_controller_by_id(controller_id)
 
         if controller is None:
-            return None
+            raise ControllerNotFound
+
+        if controller["user_id"] != user_id:
+            raise PermissionForControllerDenied
 
         client = AsyncModbusTcpClient(
             host=str(controller["ip_address"]),
@@ -44,11 +58,19 @@ class ControllerService:
             client.close()
         return result.registers
 
-    async def write_controller_data(self, controller_id: int, data_address: int = 0, data: int = 0):
+    async def write_controller_data(
+            self,
+            user_id: int,
+            controller_id: int,
+            data_address: int = 0,
+            data: int = 0
+    ):
         controller = await self.get_controller_by_id(controller_id)
-
         if controller is None:
-            return None
+            raise ControllerNotFound
+
+        if controller["user_id"] != user_id:
+            raise PermissionForControllerDenied
 
         client = AsyncModbusTcpClient(
             host=str(controller["ip_address"]),
@@ -72,9 +94,10 @@ class ControllerService:
 
         return controllers
 
-    async def create_controller(self, controller: ControllerCreateUpdate):
+    async def create_controller(self, user_id: int, controller: ControllerCreateUpdate):
         controller = Controller(**controller.model_dump())
         controller.created_at = controller.updated_at = datetime.now()
+        controller.user_id = user_id
         self.session.add(controller)
 
         await self.session.commit()
@@ -82,13 +105,21 @@ class ControllerService:
 
         return controller.to_dict()
 
-    async def update_controller(self, controller_id: int, controller_schema: ControllerCreateUpdate):
+    async def update_controller(
+            self,
+            user_id: int,
+            controller_id: int,
+            controller_schema: ControllerCreateUpdate
+    ):
         stmt = select(Controller).where(Controller.id == controller_id)
         result = await self.session.execute(stmt)
         controller = result.scalars().first()
 
         if controller is None:
-            return None
+            raise ControllerNotFound
+
+        if controller.user_id != user_id:
+            raise PermissionForControllerDenied
 
         stmt = update(Controller).where(Controller.id == controller_id).values(
             **controller_schema.model_dump()
@@ -101,15 +132,16 @@ class ControllerService:
 
         return controller.to_dict()
 
-    async def delete_controller(self, controller_id: int) -> bool | None:
+    async def delete_controller(self, user_id: int, controller_id: int):
         stmt = select(Controller).where(Controller.id == controller_id)
         result = await self.session.execute(stmt)
         controller = result.scalars().first()
 
         if controller is None:
-            return None
+            raise ControllerNotFound
+
+        if controller.user_id != user_id:
+            raise PermissionForControllerDenied
 
         await self.session.delete(controller)
         await self.session.commit()
-
-        return True
